refactor: move tasks to get filenames to clean transcripts flow
